+++ spec/client_spec.rb.orig
@@ -2,7 +2,7 @@ require 'spec_helper'
 require 'spec/support/configuration_helper'
 require 'routemaster/client'
 require 'routemaster/client/backends/sidekiq'
-require 'routemaster/client/topic'
+require 'routemaster/topic'
 require 'webmock/rspec'
 require 'sidekiq/testing'
 require 'securerandom'
@@ -465,22 +465,28 @@ describe Routemaster::Client do
       end
 
       context 'the connection to the bus is successful' do
-          include_context 'successful connection to bus'
+        include_context 'successful connection to bus'
 
         it 'expects a collection of topics' do
           expect(perform.map(&:attributes)).to eql(expected_result)
         end
       end
 
+<<<<<<< HEAD
       context 'the connection to the bus errors' do
-          include_context 'failing connection to bus'
+        include_context 'failing connection to bus'
 
         it 'expects a collection of topics' do
-          expect { perform }.to raise_error(Routemaster::Client::ConnectionError, 'failed to connect to /topics (status: 500)')
+          expect { perform }.to raise_error(RuntimeError)
         end
+=======
+      it 'expects a collection of topics' do
+        expect { perform }.to raise_error(Routemaster::Client::ConnectionError, 'failed to connect to /topics (status: 500)')
+>>>>>>> master
       end
     end
 
+<<<<<<< HEAD
     describe '#monitor_subscriptions' do
       let(:url) { 'https://bus.example.com/subscriptions' }
       let(:perform) { subject.monitor_subscriptions }
@@ -498,7 +504,7 @@ describe Routemaster::Client do
 
         it 'expects a collection of subscriptions' do
           expect(perform.map(&:attributes)).to eql(expected_result)
-      end
+        end
       end
 
       context 'the connection to the bus errors' do
@@ -507,31 +513,34 @@ describe Routemaster::Client do
         it 'expects a collection of topics' do
           expect { perform }.to raise_error(RuntimeError)
         end
-      end
-    end
-  end
-  
+=======
   describe '#reset_connection' do
+
     context 'can reset class vars to change params' do
+
       let(:instance_uuid) { SecureRandom.uuid }
+
       let(:options) {{
-          url:        'https://@bus.example.com',
-          uuid:       instance_uuid,
-          verify_ssl: false,
-          lazy: true
+        url:        'https://@bus.example.com',
+        uuid:       instance_uuid,
+        verify_ssl: false,
+        lazy: true
       }}
-      
+
       before do
-          Routemaster::Client::Connection.reset_connection
-          @stub = stub_request(:get, 'https://@bus.example.com/topics').with({basic_auth: [instance_uuid, 'x']})
+        Routemaster::Client::Connection.reset_connection
+        @stub = stub_request(:get, 'https://@bus.example.com/topics').with({basic_auth: [instance_uuid, 'x']})
           .to_return(status: 200, body: [{ name: "topic.name", publisher: "topic.publisher", events: "topic.get_count" }].to_json)
       end
-      
-      after { Routemaster::Client::Connection.reset_connection }
-      
+
+      after do
+        Routemaster::Client::Connection.reset_connection
+      end
+
       it 'connects with new params' do
-          subject.monitor_topics
-          expect(@stub).to have_been_requested
+        subject.monitor_topics
+        expect(@stub).to have_been_requested
+>>>>>>> master
       end
     end
   end
